Tratando excessoes das entidades para objeto não encontrado

+++ src/main/java/com/stefanini/servico/EnderecoServico.java
@@ -4,6 +4,7 @@ import com.stefanini.dao.EnderecoDao;
 import com.stefanini.dao.PessoaDao;
 import com.stefanini.model.Endereco;
 import com.stefanini.model.Pessoa;
+import com.stefanini.servico.exceptions.ObjectNotFoundException;
 import com.stefanini.util.IGenericService;
 
 import javax.ejb.*;
@@ -54,7 +55,8 @@ public class EnderecoServico implements Serializable {
 		}
 
 		if(endereco == null){
-			throw new BadRequestException("Não foi possível encontrar a pessoa de ID: " + entity.getIdPessoa());
+			throw new ObjectNotFoundException("Objeto não encontrado! Id: " + entity.getIdPessoa()
+					+ ", Tipo: " + Pessoa.class.getName());
 		}
 
 		return endereco;
@@ -65,7 +67,25 @@ public class EnderecoServico implements Serializable {
 	 */
 	@TransactionAttribute(TransactionAttributeType.REQUIRES_NEW)
 	public Endereco atualizar(@Valid Endereco entity) {
-		return dao.atualizar(entity);
+		List<Pessoa> pessoas = pessoaDao.getList().get();
+		Endereco endereco = null;
+
+		if(entity.getIdPessoa() == null){
+			throw new BadRequestException("Não é possível atualizar endereço sem pessoa!");
+		}
+
+		for (Pessoa pessoa: pessoas) {
+			if(pessoa.getId() == entity.getIdPessoa()){
+				endereco = dao.atualizar(entity);
+			}
+		}
+
+		if(endereco == null){
+			throw new ObjectNotFoundException("Objeto não encontrado! Id: " + entity.getIdPessoa()
+					+ ", Tipo: " + Pessoa.class.getName());
+		}
+
+		return endereco;
 	}
 
 	/**
@@ -73,7 +93,13 @@ public class EnderecoServico implements Serializable {
 	 */
 	@TransactionAttribute(TransactionAttributeType.REQUIRES_NEW)
 	public void remover(Long id) {
-		dao.remover(id);
+		Optional<Endereco> endereco = dao.encontrar(id);
+		if(endereco.isPresent()){
+			dao.remover(id);
+		} else {
+			throw new ObjectNotFoundException("Objeto não encontrado! Id: " + id
+					+ ", Tipo: " + Endereco.class.getName());
+		}
 	}
 
 	/**
@@ -89,6 +115,13 @@ public class EnderecoServico implements Serializable {
 	 */
 	@TransactionAttribute(TransactionAttributeType.REQUIRES_NEW)
 	public Optional<Endereco> encontrar(Long id) {
+		Optional<Endereco> endereco = dao.encontrar(id);
+
+		if(!endereco.isPresent()){
+			throw new ObjectNotFoundException("Objeto não encontrado! Id: " + id
+					+ ", Tipo: " + Endereco.class.getName());
+		}
+
 		return dao.encontrar(id);
 	}
 }
