+++ src/main/java/com/stefanini/servico/PerfilServico.java
@@ -1,12 +1,16 @@
 package com.stefanini.servico;
 
 import com.stefanini.dao.PerfilDao;
+import com.stefanini.model.Endereco;
 import com.stefanini.model.Perfil;
+import com.stefanini.model.Pessoa;
+import com.stefanini.servico.exceptions.ObjectNotFoundException;
 import com.stefanini.util.IGenericService;
 
 import javax.ejb.*;
 import javax.inject.Inject;
 import javax.validation.Valid;
+import javax.ws.rs.BadRequestException;
 import java.io.Serializable;
 import java.util.List;
 import java.util.Optional;
@@ -49,7 +53,12 @@ public class PerfilServico implements Serializable {
 	 */
 	@TransactionAttribute(TransactionAttributeType.REQUIRES_NEW)
 	public void remover(Long id) {
-		dao.remover(id);
+		Optional<Perfil> perfil = dao.encontrar(id);
+		if(perfil.isPresent()){
+			dao.remover(id);
+		} else {
+			throw new BadRequestException("Perfil de ID: " + id + " não encontrado!");
+		}
 	}
 
 	/**
@@ -65,6 +74,12 @@ public class PerfilServico implements Serializable {
 	 */
 	@TransactionAttribute(TransactionAttributeType.REQUIRES_NEW)
 	public Optional<Perfil> encontrar(Long id) {
-		return dao.encontrar(id);
+		Optional<Perfil> perfil = dao.encontrar(id);
+
+		if(!perfil.isPresent()){
+			throw new ObjectNotFoundException("Objeto não encontrado! Id: " + id
+					+ ", Tipo: " + Perfil.class.getName());
+		}
+		return perfil;
 	}
 }
